Function that prints fibonacci numbers less ones that start with 1 or 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i;
-	long t1, t2, nextTerm;
+	unsigned int t1, t2, nextTerm;
 
 	t1 = 0;
 	t2 = 1;
